finish corp auth and contact callback

+++ biz/src/main/java/com/dingtalk/isv/access/biz/corp/service/impl/CorpCallbackQueueServiceImpl.java
@@ -20,6 +20,8 @@ import com.dingtalk.isv.access.biz.corp.service.CorpLockService;
 import com.dingtalk.isv.common.code.ServiceResultCode;
 import com.dingtalk.isv.common.log.format.LogFormatter;
 import com.dingtalk.isv.common.model.ServiceResult;
+import com.dingtalk.isv.rsq.biz.model.RsqUser;
+import com.dingtalk.isv.rsq.biz.service.RsqCorpCallbackService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -45,6 +47,8 @@ public class CorpCallbackQueueServiceImpl implements CorpCallbackQueueService {
     private StaffManageService staffManageService;
     @Autowired
     private DeptManageService deptManageService;
+    @Autowired
+    private RsqCorpCallbackService rsqCorpCallbackService;
 
     @Override
     public ServiceResult<Void> saveCorpCallbackQueue(JSONObject jsonObject, String suiteKey) {
@@ -121,11 +125,22 @@ public class CorpCallbackQueueServiceImpl implements CorpCallbackQueueService {
         ));
         try {
             for(int i = 0; i < userIdArray.length; i ++){
-                //TODO 先同步日事清
-
-                ServiceResult<Void> fetchSr = staffManageService.getAndSaveStaff(userIdArray[i], corpId, suiteKey);
+                String userId = userIdArray[i];
+                //1 先读取钉钉并保存本地
+                ServiceResult<Void> fetchSr = staffManageService.getAndSaveStaff(userId, corpId, suiteKey);
                 if(!fetchSr.isSuccess()){
-                    return ServiceResult.failure(ServiceResultCode.SYS_ERROR.getErrCode(),ServiceResultCode.SYS_ERROR.getErrMsg());
+                    return ServiceResult.failure(fetchSr.getCode(),fetchSr.getMessage());
+                }
+                ServiceResult<StaffVO> staffVOSr = staffManageService.getStaffByCorpIdAndUserId(corpId, userId);
+                if(!staffVOSr.isSuccess()){
+                    return ServiceResult.failure(staffVOSr.getCode(), staffVOSr.getMessage());
+                }
+                StaffVO staffVO = staffVOSr.getResult();
+
+                //2 同步日事清
+                ServiceResult<StaffVO> rsqUserSr = rsqCorpCallbackService.createRsqTeamStaff(suiteKey, staffVO);
+                if(!rsqUserSr.isSuccess()){
+                    return ServiceResult.failure(rsqUserSr.getCode(),rsqUserSr.getMessage());
                 }
             }
 
@@ -157,13 +172,23 @@ public class CorpCallbackQueueServiceImpl implements CorpCallbackQueueService {
         ));
         try {
             for(int i = 0; i < userIdArray.length; i ++){
-                //TODO 先同步日事清
-
-                ServiceResult<Void> fetchSr = staffManageService.getAndSaveStaff(userIdArray[i], corpId, suiteKey);
-                if(!fetchSr.isSuccess()){
-                    return ServiceResult.failure(ServiceResultCode.SYS_ERROR.getErrCode(),ServiceResultCode.SYS_ERROR.getErrMsg());
+                String userId = userIdArray[i];
+                //1 先读取钉钉
+                ServiceResult<StaffVO> staffVOSr = staffManageService.getStaff(userId, corpId, suiteKey);
+                if(!staffVOSr.isSuccess()){
+                    return ServiceResult.failure(staffVOSr.getCode(), staffVOSr.getMessage());
                 }
-
+                StaffVO staffVO = staffVOSr.getResult();
+                //2 同步日事清
+                ServiceResult<Void> rsqUserSr = rsqCorpCallbackService.updateRsqTeamStaff(suiteKey, staffVO);
+                if(!rsqUserSr.isSuccess()){
+                    return ServiceResult.failure(rsqUserSr.getCode(),rsqUserSr.getMessage());
+                }
+                //3 保存本地
+                ServiceResult<Void> fetchSr = staffManageService.saveOrUpdateCorpStaff(staffVO);
+                if(!fetchSr.isSuccess()){
+                    return ServiceResult.failure(fetchSr.getCode(),fetchSr.getMessage());
+                }
             }
 
             return ServiceResult.success(null);
@@ -194,11 +219,26 @@ public class CorpCallbackQueueServiceImpl implements CorpCallbackQueueService {
         ));
         try {
             for(int i = 0; i < userIdArray.length; i ++){
-                //TODO 先同步日事清
-
+                String userId = userIdArray[i];
+                //1 从本地读取user
+                ServiceResult<StaffVO> staffVOSr = staffManageService.getStaffByCorpIdAndUserId(corpId, userId);
+                if(!staffVOSr.isSuccess()){
+                    return ServiceResult.failure(staffVOSr.getCode(), staffVOSr.getMessage());
+                }
+                StaffVO staffVO = staffVOSr.getResult();
+                // 如果数据库中未找到，则不做处理
+                if(null == staffVO){
+                    continue;
+                }
+                //2 同步日事清
+                ServiceResult<Void> rsqUserSr = rsqCorpCallbackService.removeUserFromTeam(suiteKey, staffVO);
+                if(!rsqUserSr.isSuccess()){
+                    return ServiceResult.failure(rsqUserSr.getCode(),rsqUserSr.getMessage());
+                }
+                //3 本地删除
                 ServiceResult<Void> delSr = staffManageService.deleteStaffByCorpIdAndUserId(corpId, userIdArray[i]);
                 if(!delSr.isSuccess()){
-                    return ServiceResult.failure(ServiceResultCode.SYS_ERROR.getErrCode(),ServiceResultCode.SYS_ERROR.getErrMsg());
+                    return ServiceResult.failure(delSr.getCode(),delSr.getMessage());
                 }
             }
 
@@ -230,13 +270,23 @@ public class CorpCallbackQueueServiceImpl implements CorpCallbackQueueService {
         ));
         try {
             for(int i = 0; i < userIdArray.length; i ++){
-                //TODO 先同步日事清
-
-                ServiceResult<Void> fetchSr = staffManageService.getAndSaveStaff(userIdArray[i], corpId, suiteKey);
-                if(!fetchSr.isSuccess()){
-                    return ServiceResult.failure(ServiceResultCode.SYS_ERROR.getErrCode(),ServiceResultCode.SYS_ERROR.getErrMsg());
+                String userId = userIdArray[i];
+                //1 先读取钉钉
+                ServiceResult<StaffVO> staffVOSr = staffManageService.getStaff(userId, corpId, suiteKey);
+                if(!staffVOSr.isSuccess()){
+                    return ServiceResult.failure(staffVOSr.getCode(), staffVOSr.getMessage());
                 }
-
+                StaffVO staffVO = staffVOSr.getResult();
+                //2 同步日事清
+                ServiceResult<Void> rsqUserSr = rsqCorpCallbackService.setRsqTeamStaffAdmin(suiteKey, staffVO);
+                if(!rsqUserSr.isSuccess()){
+                    return ServiceResult.failure(rsqUserSr.getCode(),rsqUserSr.getMessage());
+                }
+                //3 保存本地
+                ServiceResult<Void> fetchSr = staffManageService.saveOrUpdateCorpStaff(staffVO);
+                if(!fetchSr.isSuccess()){
+                    return ServiceResult.failure(fetchSr.getCode(),fetchSr.getMessage());
+                }
             }
 
             return ServiceResult.success(null);
@@ -267,13 +317,23 @@ public class CorpCallbackQueueServiceImpl implements CorpCallbackQueueService {
         ));
         try {
             for(int i = 0; i < deptIdArray.length; i ++){
-                //TODO 先同步日事清
-
-                ServiceResult<Void> fetchSr = deptManageService.getAndSaveDepartment(deptIdArray[i], corpId, suiteKey);
+                Long deptId = deptIdArray[i];
+                //1 先读取钉钉并保存本地
+                ServiceResult<Void> fetchSr = deptManageService.getAndSaveDepartment(deptId, corpId, suiteKey);
                 if(!fetchSr.isSuccess()){
-                    return ServiceResult.failure(ServiceResultCode.SYS_ERROR.getErrCode(),ServiceResultCode.SYS_ERROR.getErrMsg());
+                    return ServiceResult.failure(fetchSr.getCode(),fetchSr.getMessage());
                 }
+                ServiceResult<DepartmentVO> deptVOSr = deptManageService.getDepartmentByCorpIdAndDeptId(corpId, deptId);
+                if(!deptVOSr.isSuccess()){
+                    return ServiceResult.failure(deptVOSr.getCode(), deptVOSr.getMessage());
+                }
+                DepartmentVO deptVO = deptVOSr.getResult();
 
+                //2 同步日事清
+                ServiceResult<DepartmentVO> rsqDeptSr = rsqCorpCallbackService.createRsqDepartment(suiteKey, deptVO);
+                if(!rsqDeptSr.isSuccess()){
+                    return ServiceResult.failure(rsqDeptSr.getCode(),rsqDeptSr.getMessage());
+                }
             }
 
             return ServiceResult.success(null);
@@ -304,13 +364,23 @@ public class CorpCallbackQueueServiceImpl implements CorpCallbackQueueService {
         ));
         try {
             for(int i = 0; i < deptIdArray.length; i ++){
-                //TODO 先同步日事清
-
-                ServiceResult<Void> fetchSr = deptManageService.getAndSaveDepartment(deptIdArray[i], corpId, suiteKey);
-                if(!fetchSr.isSuccess()){
-                    return ServiceResult.failure(ServiceResultCode.SYS_ERROR.getErrCode(),ServiceResultCode.SYS_ERROR.getErrMsg());
+                Long deptId = deptIdArray[i];
+                //1 先读取钉钉
+                ServiceResult<DepartmentVO> deptVOSr = deptManageService.getDept(deptId, corpId, suiteKey);
+                if(!deptVOSr.isSuccess()){
+                    return ServiceResult.failure(deptVOSr.getCode(), deptVOSr.getMessage());
                 }
-
+                DepartmentVO deptVO = deptVOSr.getResult();
+                //2 同步日事清
+                ServiceResult<DepartmentVO> rsqDeptSr = rsqCorpCallbackService.updateRsqDepartment(suiteKey, deptVO);
+                if(!rsqDeptSr.isSuccess()){
+                    return ServiceResult.failure(rsqDeptSr.getCode(),rsqDeptSr.getMessage());
+                }
+                //3 保存本地
+                ServiceResult<Void> fetchSr = deptManageService.saveOrUpdateCorpDepartment(deptVO);
+                if(!fetchSr.isSuccess()){
+                    return ServiceResult.failure(fetchSr.getCode(),fetchSr.getMessage());
+                }
             }
 
             return ServiceResult.success(null);
@@ -341,13 +411,27 @@ public class CorpCallbackQueueServiceImpl implements CorpCallbackQueueService {
         ));
         try {
             for(int i = 0; i < deptIdArray.length; i ++){
-                //TODO 先同步日事清
-
-                ServiceResult<Void> fetchSr = deptManageService.deleteStaffByCorpIdAndUserId(corpId, deptIdArray[i]);
-                if(!fetchSr.isSuccess()){
-                    return ServiceResult.failure(ServiceResultCode.SYS_ERROR.getErrCode(),ServiceResultCode.SYS_ERROR.getErrMsg());
+                Long deptId = deptIdArray[i];
+                //1 从本地读取user
+                ServiceResult<DepartmentVO> deptVOSr = deptManageService.getDepartmentByCorpIdAndDeptId(corpId, deptId);
+                if(!deptVOSr.isSuccess()){
+                    return ServiceResult.failure(deptVOSr.getCode(), deptVOSr.getMessage());
                 }
-
+                DepartmentVO deptVO = deptVOSr.getResult();
+                // 如果数据库中未找到，则不做处理
+                if(null == deptVO){
+                    continue;
+                }
+                //2 同步日事清
+                ServiceResult<Void> rsqDeptSr = rsqCorpCallbackService.deleteRsqDepartment(suiteKey, deptVO);
+                if(!rsqDeptSr.isSuccess()){
+                    return ServiceResult.failure(rsqDeptSr.getCode(),rsqDeptSr.getMessage());
+                }
+                //3 本地删除
+                ServiceResult<Void> delSr = deptManageService.deleteDepartmentByCorpIdAndUserId(corpId, deptId);
+                if(!delSr.isSuccess()){
+                    return ServiceResult.failure(delSr.getCode(),delSr.getMessage());
+                }
             }
 
             return ServiceResult.success(null);
